requirements.txt
asgiref==3.6.0
cloudinary==1.33.0
dj-database-url==2.0.0
dj-static==0.0.6
Django==3.2.13
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
static3==0.7.0
urllib3==1.26.16
whitenoise==6.4.0
